Split RCN class and RCN rules tests
